Small fixes

+++ src/simulation.rs
@@ -1,2 +1,2 @@
-pub const G: f32 = 6.6743e-11;
-pub const AU: f32 = 150e9;
+pub const G: f64 = 6.6743e-11;
+pub const AU: f64 = 150e9;
